requirements.txt
python-binance==1.0.19
matplotlib==3.8.2
pandas==2.1.4

